requirements.txt
django
djangorestframework
python-decouple
black
pyotp
djangorestframework-simplejwt==5.2.2
